feat: move snapshot implementation into @vitest/snapshot
