+++ packages/snapshot/src/environment.ts
@@ -1,0 +1,2 @@
+export { NodeSnapshotEnvironment } from './env/node'
+export type { SnapshotEnvironment } from './types/environment'
