+++ packages/utils/src/index.ts
@@ -8,3 +8,4 @@ export * from './constants'
 export * from './colors'
 export * from './error'
 export * from './descriptors'
+export * from './source-map'
