+++ packages/vitest/package.json
@@ -139,6 +139,7 @@
     "@types/node": "*",
     "@vitest/expect": "workspace:*",
     "@vitest/runner": "workspace:*",
+    "@vitest/snapshot": "workspace:^0.29.3",
     "@vitest/spy": "workspace:*",
     "@vitest/utils": "workspace:*",
     "acorn": "^8.8.1",
@@ -169,7 +170,6 @@
     "@types/istanbul-reports": "^3.0.1",
     "@types/jsdom": "^20.0.1",
     "@types/micromatch": "^4.0.2",
-    "@types/natural-compare": "^1.4.1",
     "@types/prompts": "^2.4.2",
     "@types/sinonjs__fake-timers": "^8.1.2",
     "birpc": "^0.2.3",
@@ -185,10 +185,8 @@
     "happy-dom": "^8.3.2",
     "jsdom": "^21.1.0",
     "log-update": "^5.0.1",
-    "magic-string": "^0.27.0",
     "micromatch": "^4.0.5",
     "mlly": "^1.1.0",
-    "natural-compare": "^1.4.0",
     "p-limit": "^4.0.0",
     "pkg-types": "^1.0.1",
     "playwright": "^1.28.0",
