+++ packages/vitest/src/browser.ts
@@ -1,4 +1,3 @@
 export { startTests } from '@vitest/runner'
 export { setupCommonEnv } from './runtime/setup.common'
-export { setupSnapshotEnvironment } from './integrations/snapshot/env'
 export { takeCoverageInsideWorker, stopCoverageInsideWorker, getCoverageProvider, startCoverageInsideWorker } from './integrations/coverage'
