+++ packages/vitest/src/index.ts
@@ -15,7 +15,7 @@ export { runOnce, isFirstRun } from './integrations/run-once'
 export * from './integrations/chai'
 export * from './integrations/vi'
 export * from './integrations/utils'
-export type { SnapshotEnvironment } from './integrations/snapshot/env'
+export type { SnapshotEnvironment } from '@vitest/snapshot/environment'
 
 export * from './types'
 export * from './api/types'
