+++ packages/vitest/src/runtime/setup.node.ts
@@ -4,12 +4,11 @@ import { installSourcemapsSupport } from 'vite-node/source-map'
 import { createColors, setupColors } from '@vitest/utils'
 import { environments } from '../integrations/env'
 import type { Environment, ResolvedConfig } from '../types'
+import { VitestSnapshotEnvironment } from '../integrations/snapshot/environments/node'
 import { getSafeTimers, getWorkerState } from '../utils'
 import * as VitestIndex from '../index'
 import { RealDate } from '../integrations/mock/date'
 import { expect } from '../integrations/chai'
-import { setupSnapshotEnvironment } from '../integrations/snapshot/env'
-import { NodeSnapshotEnvironment } from '../integrations/snapshot/environments/node'
 import { rpc } from './rpc'
 import { setupCommonEnv } from './setup.common'
 import type { VitestExecutor } from './execute'
@@ -24,11 +23,15 @@ export async function setupGlobalEnv(config: ResolvedConfig) {
     enumerable: false,
   })
 
+  const state = getWorkerState()
+
+  if (!state.config.snapshotOptions.snapshotEnvironment)
+    state.config.snapshotOptions.snapshotEnvironment = new VitestSnapshotEnvironment()
+
   if (globalSetup)
     return
 
   globalSetup = true
-  setupSnapshotEnvironment(new NodeSnapshotEnvironment())
   setupColors(createColors(isatty(1)))
 
   const _require = createRequire(import.meta.url)
@@ -36,8 +39,6 @@ export async function setupGlobalEnv(config: ResolvedConfig) {
   _require.extensions['.css'] = () => ({})
   _require.extensions['.scss'] = () => ({})
   _require.extensions['.sass'] = () => ({})
-
-  const state = getWorkerState()
 
   installSourcemapsSupport({
     getSourceMap: source => state.moduleCache.getSourceMap(source),
